Versão Funfando

+++ Vendas/src/br/com/exemplo/vendas/apresentacao/actions/compra/InserirCompraACT.java
@@ -22,7 +22,7 @@ public class InserirCompraACT implements Action
 			throws LayerException
 	{
 		
-		BigInteger numero = new BigInteger(request.getParameter("numero"));
+		Integer numero = new Integer(request.getParameter("numero"));
 		Date data = new Date(request.getParameter("data"));
 		String responsavel = request.getParameter("responsavel");
 		String situacao = request.getParameter("situacao");
