+++ Vendas/src/br/com/exemplo/vendas/negocio/interfaces/CompraInterface.java
@@ -21,6 +21,6 @@ public interface CompraInterface
 			throws br.com.exemplo.vendas.util.exception.LayerException, java.rmi.RemoteException ;
 
 	public br.com.exemplo.vendas.util.dto.ServiceDTO getCompra(
-			br.com.exemplo.vendas.util.dto.ServiceDTO requestDTO, BigInteger codigo )
+			br.com.exemplo.vendas.util.dto.ServiceDTO requestDTO, Integer codigo )
 			throws br.com.exemplo.vendas.util.exception.LayerException, java.rmi.RemoteException ;
 }
